+++ Config/widgets_map.php
@@ -1,5 +1,6 @@
 <?php
 $config = array('WidgetsMap' => array(
+	'MeCms.albums'			=> 'MeCms.PhotosWidget::albums',
 	'MeCms.categories'		=> 'MeCms.PostsWidget::categories',
 	'MeCms.latest_posts'	=> 'MeCms.PostsWidget::latest',
 	'MeCms.pages'			=> 'MeCms.PagesWidget::pages',